+++ src/main/java/com/example/SOPSbackend/service/DoctorService.java
@@ -3,7 +3,7 @@ package com.example.SOPSbackend.service;
 import com.example.SOPSbackend.exception.InternalValidationException;
 import com.example.SOPSbackend.model.DoctorEntity;
 import com.example.SOPSbackend.model.VaccinationSlotEntity;
-import com.example.SOPSbackend.model.converter.VaccinationSlotStatus;
+import com.example.SOPSbackend.model.converter.ResponseDictionary;
 import com.example.SOPSbackend.repository.DoctorRepository;
 import com.example.SOPSbackend.repository.VaccinationRepository;
 import com.example.SOPSbackend.repository.VaccinationSlotRepository;
@@ -57,7 +57,7 @@ public class DoctorService {
         vaccinationSlotRepository.deleteByIdAndDoctor(id, doctor);
     }
 
-    public Page<VaccinationSlotStatus> getVaccinationSlots(DoctorEntity doctor, int page, String startDate, String  endDate, String onlyReserved) {
+    public Page<ResponseDictionary> getVaccinationSlots(DoctorEntity doctor, int page, String startDate, String  endDate, String onlyReserved) {
         Pageable thisPage = PageRequest.of(page - 1, ITEMS_PER_PAGE, Sort.by("date"));
         Page<VaccinationSlotEntity> mySlots;
         LocalDateTime sDate, eDate;
@@ -89,7 +89,7 @@ public class DoctorService {
         }
 
         var myVaccinations = vaccinationRepository.findMatchingVaccinations(mySlots.toList());
-        return mySlots.map((vs) -> new VaccinationSlotStatus(vs,
+        return mySlots.map((vs) -> new ResponseDictionary(vs,
                         myVaccinations.stream()
                                       .filter((v) -> v.getVaccinationSlot().equals(vs))
                                       .findAny()
